+++ quantforge/__init__.py
@@ -312,6 +312,7 @@ from .scale_tests import ansari_bradley_test, mood_test
 from .kruskal_wallis import kruskal_wallis_test, friedman_test
 from .dunn_test import dunn_test
 from .jonckheere import jonckheere_terpstra_test
+from .cochran_mcnemar import mcnemar_test, cochran_q_test
 from .kalman_beta import kalman_regression_beta
 from .theil_sen import theil_sen
 from .robust_stats import (
@@ -871,7 +872,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.885.0"
+__version__ = "1.886.0"
 
 __all__ = [
     "OptionType",
@@ -1349,6 +1350,8 @@ __all__ = [
     "friedman_test",
     "dunn_test",
     "jonckheere_terpstra_test",
+    "mcnemar_test",
+    "cochran_q_test",
     "kalman_regression_beta",
     "theil_sen",
     "median_absolute_deviation",